# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My Fsirst API",
        "version": "0.1"
    },
    "paths": {
...
        "/api/boards/board-gallery/{tid}": {
            "get": {
                "tags": [
                    "게시판"
                ],
                "description": "팀 갤러리 사진 리스트",
                "operationId": "7c3dd583e4326428b7efb52f2b9af98b",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "500": {
                        "description": "Fail"
                    }
                }
            },
            "post": {
                "tags": [
                    "게시판"
                ],
                "description": "팀 갤러리 사진 생성",
                "operationId": "f0fd7af58803f36991231d6f16fcd4a8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "tid": {
                                        "description": "팀아이디",
                                        "type": "string",
                                        "example": "1"
                                    },
                                    "files[]": {
                                        "description": "이미지 input name=files[]",
                                        "type": "file",
                                        "example": "file.jpg"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "500": {
                        "description": "Fail"
                    }
                }
            }
        },
        "/api/boards/board-notice/{tid}{per_page?}{page?}": {
            "get": {
                "tags": [
                    "게시판"
                ],
                "description": "팀 공지사항 리스트 [per_page=>페이지 당 레코드 갯수] [page =>보여줄 페이지]",
                "operationId": "7d209b331abc45241844fe90f6b24351",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "500": {
                        "description": "Fail"
                    }
                }
            }
        },
        "/api/boards/board-notice/{tid}/{sid}": {
            "get": {
                "tags": [
                    "게시판"
                ],
                "description": "팀 공지사항 상세 불러오기",
                "operationId": "5310f80bceb9bbd5a7e9d6f03fca9cc7",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "500": {
                        "description": "Fail"
                    }
                }
            },
            "post": {
                "tags": [
                    "게시판"
                ],
                "description": "팀 공지사항 게시판 수정 및 삭제",
                "operationId": "d85a7182310f75209bbaa30698520da0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "description": "제목",
                                        "type": "string",
                                        "example": "제목"
                                    },
                                    "contents": {
                                        "description": "내용",
                                        "type": "string",
                                        "example": "내용"
                                    },
                                    "del_yn": {
                                        "description": "삭제유무",
                                        "type": "string",
                                        "example": "Y"
                                    },
                                    "files[]": {
                                        "description": "이미지 input name=files[]",
                                        "type": "file",
                                        "example": "file.jpg"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "500": {
                        "description": "Fail"
                    }
                }
            }
        },
        "/api/boards/board-notice/{tid}": {
            "post": {
                "tags": [
                    "게시판"
                ],
                "description": "팀 공지사항 생성",
                "operationId": "bdf815c25e141dad5bea342ba0904d14",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "description": "제목",
                                        "type": "string",
                                        "example": "제목"
                                    },
                                    "contents": {
                                        "description": "내용",
                                        "type": "string",
                                        "example": "내용"
                                    },
                                    "files[]": {
                                        "description": "이미지 input name=files[]",
                                        "type": "file",
                                        "example": "file.jpg"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "500": {
                        "description": "Fail"
                    }
                }
            }
        },
        "/api/boards/board-inquire/{cid}": {
            "get": {
                "tags": [
                    "게시판"
                ],
                "description": "경기 공지/문의 리스트",
                "operationId": "a737c97d28a3082b7fe86ea3ae69fcf8",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "500": {
                        "description": "Fail"
                    }
                }
            },
            "post": {
                "tags": [
                    "게시판"
                ],
                "description": "경기 공지/문의 생성",
                "operationId": "4d9b65992724ac9ac6d4319d100e4ac2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "description": "제목",
                                        "type": "string",
                                        "example": "제목"
                                    },
                                    "contents": {
                                        "description": "내용",
                                        "type": "string",
                                        "example": "내용"
                                    },
                                    "files[]": {
                                        "description": "이미지 input name=files[]",
                                        "type": "file",
                                        "example": "file.jpg"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "500": {
                        "description": "Fail"
                    }
                }
            }
        },
        "/api/boards/board-inquire/{cid}/{sid}": {
            "get": {
                "tags": [
                    "게시판"
                ],
                "description": "경기 공지/문의  상세 불러오기",
                "operationId": "bdc84b81e9ef4d24f2e95bc183b8a6b9",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "500": {
                        "description": "Fail"
                    }
                }
            }
        },
        "/api/boards/board-inquire/{tid}/{sid}": {
            "post": {
                "tags": [
                    "게시판"
                ],
                "description": "경기 공지/문의 수정 및 삭제",
                "operationId": "44a7b3f9ac58e85bcc28edec20bed95c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "description": "제목",
                                        "type": "string",
                                        "example": "제목"
                                    },
                                    "contents": {
                                        "description": "내용",
                                        "type": "string",
                                        "example": "내용"
                                    },
                                    "del_yn": {
                                        "description": "삭제유무",
                                        "type": "string",
                                        "example": "Y"
                                    },
                                    "files[]": {
                                        "description": "이미지 input name=files[]",
                                        "type": "file",
                                        "example": "file.jpg"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "500": {
                        "description": "Fail"
                    }
                }
            }
        },
        "/api/boards/board-reply/{sid}": {
            "post": {
                "tags": [
                    "게시판"
                ],
                "description": "경기 공지/문의 댓글달기",
                "operationId": "b8929128d3cb92288945b8a34b111cd0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "description": "제목",
                                        "type": "string",
                                        "example": "제목"
                                    },
                                    "contents": {
                                        "description": "내용",
                                        "type": "string",
                                        "example": "내용"
                                    },
                                    "files[]": {
                                        "description": "이미지 input name=files[]",
                                        "type": "file",
                                        "example": "file.jpg"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "500": {
                        "description": "Fail"
                    }
                }
            }
        },
...
    }
}